requirements.txt
PyQt5==5.15.10
PyQtGraph==0.13.3
PyOpenGL==3.1.7
numpy>=1.26.4
